Alma (#922)
* purge items_by_bib call. update vcr cassettes
'

* One passing test example.

Clean-up Notes.

* Combo id cleanup (#930)

Co-authored-by: Shaun Ellis <shaun@sdellis.com>

* Deals with enum and chron key value changes. (#937)

* Adding in the alma gem
Change HoldItem to contact alma instead of voyager

* Add known alma status_label values. (#940)

* Updating to make an In Process Feature test work
Ended up changing the charged method to look at the status_label not the status
Updated example in requestable spec to be an acutal missing item

* renaming annexa library to annex to match with data
Updated examples in feature test to new alma examples

* Changing from enum to enum_display as the enumeration key (#944)

Also allow fill in eligible if any item is enumerated, not just the first

* rename architecture library to arch
Also double check the enum_display has conetent and not just blank

* Updating feature test to reflect alma ids and locations
Also updating 'Technical - Migration' to be borrow Direct or ILL Request, not on the shelf
Also updating illiad-transaction specs to mock the ocrrect metadata

* mock SCSB calls to the catalog
Pulled data from production to fixtures
Modified VCR setup to not call search-base for the items
Updated tests to mock with downloaded fixtures

* fixes a few tests wrt on_order items

* fixes load_items_by_mfhd method in request.rb

* fixes holding_id for a test, now passes

* fixes test for borrow direct eligible by updating record

* removes test for holdings with e-resources since this is no longer the case in alma

* fixes test by fixing holding_id

* Fix alma-updated failing tests (#938)

* fixes a few tests wrt on_order items

* fixes load_items_by_mfhd method in request.rb

* fixes holding_id for a test, now passes

* fixes test for borrow direct eligible by updating record

* removes test for holdings with e-resources since this is no longer the case in alma

* fixes test by fixing holding_id

* Updating the cassettes so circle ci passes (#952)

* Updating scsb feature tests to pass
Also changing temporary location to utilize the code

* Update typo in Acq status. (#958)

* Updating to not error out for onlin only items

* Updating helper tests to pass

* fixes several tests by updating holding_id for Aeon requestable without a barcode

* fixes on_order tests by updating factory with test ids

* Updating request model test to use current on order example
fixes #928

* Fixing the on order feature tests

* fixes a few tests by updating to alma location labels

* Updated a few request model tests to pass
Also deleted the request model cassette and rerecorded it.  That way it removes all the old data and only keeps the alma data

* fixes order_pending test failures

* reverses the order_pending updates; updates on_hold factory items

* updates thesis bib id to get passing tests

* Adding refreshed thesis fixture to cassettes. (#968)

* Changing numismatics location to new alma location `rare$num`
Changed to stub test vs using vcr recording
Fix a couple other request tests

* Moving thesis catalog records for alma
Changed to use a fixture instead of VCRed data
fixes #956

* Adding the thesis fixtures that were forgotten

* Changing clancy base to come from config instead of environment variable
Also fixing the submission specs

* Fixes (#973)

* Fix test specs, align items with Alma status labels.

* Remove redundant on order specs (no more pending orders in alma).

* Comment out specs related to multi-holding checks we no longer consider.

* Comment out etas.

* return to on shelf services.

* Fix aeon openrul

* Move in process/on order statuses to unavailable

* Available should look at status label.

* Change pickup information to json to include GFA and Alma code
Also default pick_up_location_code to firestone if it is missing

* Updating the requestable specs for no items and on order
No items are assumed unavailable under Alma
On order is pending until there is a bug fix in bibdata

* Removing on order and in process from unavailable status
On order and in process are not available, but they are not entirely unavailable.  The are special cases that should be handled differently.
Also removed the cassette and rebuilt it now that all test are passing

* Changing from looking at library location to remote_storage location for recap check

* Adding a Alma hold request for Princeton owned ReCap items
Only do it for requests where the book leaves ReCap

Move the hold call to a separate call from the initialized to allow for calling it from ReCAP

* Update on order/in process feature specs. (#984)

* Add campus access info (#983)

* Updating to allow borrow direct and ILL for Resource shared items

* Removes Copy 0, fixes #i986

* adds spec test for helper.enum_copy_display

* fixes rubocop failure

* updates application helper to show additional enumeration data even when copy is zero

* fixes rubocop complaints

* fixes broken spec

* Extending the Requests::Patron#current_patron method to handle 429 HTTP responses from the BibData endpoint

* Simplifying the Requests::Patron#api_request_patron case/when statements

* Removing the deprecated custom error classes from the Requests::Patron Model

* Catching when no map error when no locationns are present

* Fixing the build from the rebase

* fixes #946

* Change engineering to engineer to match bibdata

* Changing where the pick_up_code comes from
Previously we assumed that there would be a field pick_up_code, but now we realize we should be getting it from the library code

* Updating recap workflow to send an email on Alma hold failure
If the SCSB request works, but the alma hold request fails the user will believe everything went ok.  We will just send an email to recap problems to let them know the hold failed to generate

* Updating item to use scsb status only for scsb items
Needed to make scsb? method work correctly by sending in the location code to the item

* Changing the color check from charged to the status label
Available is always green and Not Available is always red to match with the catalog

* Changing the offsite location to be recap
Now that some items that are stored in recap are now showing their library (like firestone) instead of recap for thier libray codes we need to just respond with the off site location of recap if the item is out at recap

* Default status to In Process for ReCAP items with no status in ReCAP
Updated the location information in a request to be singular, since we are only dealing with one mfhd (or location) now.
Availability was not getting queried for the new locations where the library code was not equal to recap, so we updated the check in the request model
Having Recap availability being queried change a number of tests

* Fixing the double status
Also allows the ReCAP status to fully override the Alma status by setting it to nil

* Updating Recap  items to allow Alma work order statuses to remain
Also add 'Preservation and Conservation' to unavailabel statuses

* Fixing up the tests for the ReCAP closure

* Updating tests that were new for alma to be xit

Co-authored-by: Shaun Ellis <shaun@sdellis.com>
Co-authored-by: Carolyn Cole <cac9@princeton.edu>
Co-authored-by: carolyncole <1599081+carolyncole@users.noreply.github.com>
Co-authored-by: jrgriffiniii <jrgriffiniii@gmail.com>